Cleanup the load order in synchronized object
