+++ lib/concurrent/atomic_reference/ruby.rb
@@ -1,5 +1,5 @@
 if defined? Concurrent::CAtomic
-  require 'concurrent/extension_helper'
+  require 'concurrent/native_extensions'
   require 'concurrent/atomic_reference/direct_update'
   require 'concurrent/atomic_reference/numeric_cas_wrapper'
