+++ lib/concurrent/native_extensions.rb
@@ -1,0 +1,2 @@
+require 'concurrent/native_extensions/before'
+require 'concurrent/native_extensions/load'
