+++ lib/concurrent/synchronized_object_implementations/monitor.rb
@@ -1,0 +1,21 @@
+module Concurrent
+  module SynchronizedObjectImplementations
+    class Monitor < Mutex
+      def initialize
+        @__lock__do_not_use_directly      = ::Monitor.new
+        @__condition__do_not_use_directly = @__lock__do_not_use_directly.new_cond
+      end
+
+      def synchronize
+        @__lock__do_not_use_directly.synchronize { yield }
+      end
+
+      private
+
+      def ns_wait(timeout = nil)
+        @__condition__do_not_use_directly.wait timeout
+        self
+      end
+    end
+  end
+end
